+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- CSS Handle for unavailable products update. 
+
+
 ## [3.171.0] - 2024-01-09
 
 ## [3.170.0] - 2023-11-28